requirements.txt
openai
exa-py
python-dotenv
pypdf
pandas
faiss-cpu
numpy
playwright 
time
boto3>=1.28.0